tests/wpt/tests/trusted-types/trusted-types-eval-reporting.html

<!DOCTYPE html>
<head>
  <script nonce="123" src="/resources/testharness.js"></script>
  <script nonce="123" src="/resources/testharnessreport.js"></script>
  <script nonce="123" src="./support/csp-violations.js"></script>
</head>
<body>
  <script nonce="123">
  // CSP insists the "trusted-types: ..." directives are delivered as headers
  // (rather than as "<meta http-equiv" tags). This test assumes the following
  // headers are set in the .headers file:
  //
  //   Content-Security-Policy: script-src 'unsafe-inline' 'unsafe-eval'; report-uri ...
  //   Content-Security-Policy: connect-src 'none'
  //   Content-Security-Policy: require-trusted-types-for 'script'
  //
  // The last rule is there so we can provoke a CSP violation report at will.
  // The intent is that in order to test that a violation has *not* been thrown
  // (and without resorting to abominations like timeouts), we force a *another*
  // CSP violation (by violating the img-src rule) and when that event is
  // processed we can we sure that an earlier event - if it indeed occurred -
  // must have already been processed.

  // A sample policy we use to test trustedTypes.createPolicy behaviour.
  const id = x => x;
  const a_policy = {
    createHTML: id,
    createScriptURL: id,
    createScript: id,
  };
  const scriptyPolicy = trustedTypes.createPolicy('allowEval', a_policy);

  promise_test(async t => {
    const input = 'beacon="should not run"';
    let beacon = 'never_overwritten';
    let violation = await trusted_type_violation_for(EvalError, _ =>
      eval(input)
    );
    assert_true(violation.originalPolicy.includes("require-trusted-types-for 'script'"));
    assert_equals(violation.sample, `eval|${clipSampleIfNeeded(input)}`);
    assert_equals(beacon, 'never_overwritten');
  }, "Trusted Type violation report: evaluating a string.");

  promise_test(async t => {
    let beacon = 'never_overwritten2';
    await no_trusted_type_violation_for(_ =>
      eval(scriptyPolicy.createScript('beacon="i ran"'))
    );
    assert_equals(beacon, 'i ran');
  }, "Trusted Type violation report: evaluating a Trusted Script.");

  </script>
</body>
